RECURSION
• Recursion is the process of repeating items in a self-similar way.
• Recursion is the process which comes into existence when a
function calls a copy of itself to work on a smaller problem.
• Any function which calls itself is called recursive function, and
such function calls are called recursive calls.
• A recursive function performs the tasks by dividing it into the
subtasks. There is a termination condition defined in the
function which is satisfied by some specific subtask. After this,
the recursion stops and the final result is returned from the
function.
• it is often done in such a way that no infinite loop or infinite chain
of references can occur.