PUSH / POP
A stack is an abstract structure containing a set of homogeneous components and is based on the principle of Last In First Out (LIFO). It is a usually used abstract data type with two major operations, namely push and pop. Push and pop are carried out on the topmost element, which is the item best recently added to the stack. The data structure stack denotes an Abstract Data Type (ADT), generally used in several programming languages. For example: A pile of plates or a deck of cards.

The stack is generally used in converting and assessing expressions in Polish notations, i.e. Infix, Prefix Postfix. In case of arrays and linked lists, these two agree programmers to insert and delete components from any place within the list, i.e., from the beginning or the end or even from the middle also. But in computer programming and expansion, there may raise some situations where insertion and deletion want only at one end decline at the beginning or end of the list. The stack is a linear data structure, and all the insertion and deletion of its values are complete in the same end which is called the top of the stack. Let us assume, take the real-life example of a stack of plates or a pile of books etc. As the item in this form of data structure can be detached or added from the top only which means the last item to be added to the stack is the first item to be removed. So you can say that the stack follows the Last In First Out (LIFO) structure.